Fix missing image in the doc build
The docs/conf.py script works under the assumption of the README.rst is
in the docs/ directory, not in the root.

The rst file in the docs/ directory imports the one in the root, but
that import doesn't include images.

To bridge that gap, we're copying the image from the root to the docs
directory at runtime.

At first I thought it might be better to generate the image from the dot
file into the docs, but we also need it in the root to render it in the
GitHub UI.